cleans up dependencies
